Refactoring for future functions
